[수정] 일정 장소 x,y좌표값 타입 double로 변경

+++ src/main/java/com/YH/yeohaenghama/domain/itinerary/dto/PlaceJoinDTO.java
@@ -22,9 +22,9 @@ public class PlaceJoinDTO {
     @Schema(description = "메모")
     private String memo;
     @Schema(description = "X　좌표")
-    private Integer mapx;
+    private double mapx;
     @Schema(description = "Y　좌표")
-    private Integer mapy;
+    private double mapy;
     @Data
     public static class Request {
         private Integer day;
@@ -35,7 +35,7 @@ public class PlaceJoinDTO {
         private String placeName;
         private String addr1;
         private String memo;
-        private Integer mapx;
-        private Integer mapy;
+        private double mapx;
+        private double mapy;
     }
 }
